Refactored folder structure, ui fixes
